+++ src/core/bridge.class.js
@@ -1,0 +1,29 @@
+import {
+    Template,
+    License,
+    Datz,
+    Tag,
+    Category,
+    Post
+} from './descriptive_class.js';
+import {
+    GString,
+    Collection,
+    Cache,
+    Correspond,
+    NULL_OBJECT,
+} from './struct.js';
+
+export {
+    Template,
+    License,
+    Datz,
+    Tag,
+    Category,
+    Post,
+    GString,
+    Collection,
+    Cache,
+    Correspond,
+    NULL_OBJECT
+}