+++ src/core/descriptive_class.js
@@ -1,6 +1,7 @@
 import { parse } from "marked";
+import { minify } from "html-minifier";
 import { statSync, readFileSync } from 'fs';
-import { Cache, Collection, GString } from './struct.js';
+import { Collection, GString } from './struct.js';
 import { word_count } from "./text_process.js";
 import { warn } from "./run.js";
 import db from "./database.js";
@@ -245,8 +246,8 @@ class Post {
 
         this.paths.website = `/${tempor_val}/`
         this.paths.local = `${tempor_val}/index.html`
-        this.parsed.content = parse(this.content)
-        this.parsed.foreword = parse(this.foreword)
+        this.parsed.content = minify(parse(this.content,{mangle:false,headerIds:false}),{removeComments:true,collapseWhitespace:true})
+        this.parsed.foreword = minify(parse(this.foreword,{mangle:false,headerIds:false}),{removeComments:true,collapseWhitespace:true})
     }
     setPath(path) {
         this.pathto = path;